Ej2: Modified Client & Server to take config changes
Currently both client and server have config files copied onto the
docker images. Also environment variables are defined in the
docker-compose.

To avoid having to build an image everytime we want to perform a
configuration change, I've added volumes definition for both the server
and client.

Also, the client takes environmental variables preference over the
configuration file. Therefore for any changes in the configuration file
to be taken into account, I've commented out the environmental variables
defined in the `docker-compose`